Move traits into own module
