+++ Cargo.toml
@@ -33,7 +33,7 @@ default = ["24bits", "50bits", "libm"]
 # May result in slight numerical instability, which can be reduced if the target CPU has fused multiply-add instructions.
 estrin = ["dep:fast_polynomial"]
 # Expose the `LambertW0` and `LambertWm1` traits and their implementations.
-trait = ["dep:num-traits"]
+traits = ["dep:num-traits"]
 # If the `std` feature is disabled, this feature uses the [`libm`](https://crates.io/crates/libm) crate
 # to compute square roots and logarithms instead of the standard library.
 libm = ["dep:libm", "fast_polynomial?/libm", "num-traits?/libm"]
